langchain/covid_q_a_app/requirements.txt
langchain
flask
google-cloud-aiplatform
pypdf
PyPDF2
faiss-cpu
huggingface_hub
sentence-transformers
python-dotenv
streamlit-extras
pinecone-client